+++ mycodo/mycodo_client.py
@@ -135,9 +135,12 @@ class DaemonControl:
     def refresh_conditionals(self):
         return self.rpyc_client.root.refresh_conditionals()
 
-    def trigger_conditional_actions(self, conditional_id, message=''):
+    def trigger_conditional_actions(
+            self, conditional_id, message='', output_state=None,
+            on_duration=None, duty_cycle=None):
         return self.rpyc_client.root.trigger_conditional_actions(
-            conditional_id, message)
+            conditional_id, message, output_state=output_state,
+            on_duration=on_duration, duty_cycle=duty_cycle)
 
     def terminate_daemon(self):
         return self.rpyc_client.root.terminate_daemon()