+++ mycodo/mycodo_daemon.py
@@ -220,9 +220,13 @@ def mycodo_service(mycodo):
             return mycodo.output_setup(action, relay_id)
 
         @staticmethod
-        def exposed_trigger_conditional_actions(conditional_id, message=''):
+        def exposed_trigger_conditional_actions(
+                conditional_id, message='', output_state=None,
+                on_duration=None, duty_cycle=None):
             """Return the output state (not pin but whether output is on or off"""
-            return mycodo.trigger_conditional_actions(conditional_id, message)
+            return mycodo.trigger_conditional_actions(
+                conditional_id, message, output_state=output_state,
+                on_duration=on_duration, duty_cycle=duty_cycle)
 
         @staticmethod
         def exposed_terminate_daemon():
@@ -867,10 +871,13 @@ class DaemonController:
                 "Output controller had an issue stopping: {err}".format(
                     err=err))
 
-    def trigger_conditional_actions(self, conditional_id, message=''):
+    def trigger_conditional_actions(self, conditional_id, message='',
+                                    output_state=None, on_duration=None,
+                                    duty_cycle=None):
         try:
             return self.controller['Conditional'].trigger_conditional_actions(
-                conditional_id, message=message)
+                conditional_id, message=message, output_state=output_state,
+                on_duration=on_duration, duty_cycle=duty_cycle)
         except Exception as except_msg:
             message = "Could not trigger Conditional Actions:" \
                       " {err}".format(err=except_msg)